+++ src/main/java/com/nhnacademy/authservice/adapter/UserAdapter.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.authservice.adapter;
 
-import com.nhnacademy.authservice.domain.OAuth2UserCreateRequestDto;
-import com.nhnacademy.authservice.domain.UserResponse;
+import com.nhnacademy.authservice.domain.request.OAuth2UserCreateRequestDto;
+import com.nhnacademy.authservice.domain.response.UserResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
