+++ src/main/java/com/nhnacademy/authservice/client/token/OAuth2TokenClient.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.authservice.client.token;
 
-import com.nhnacademy.authservice.domain.OAuth2TokenResponse;
+import com.nhnacademy.authservice.domain.response.OAuth2TokenResponse;
 
 public interface OAuth2TokenClient {
     OAuth2TokenResponse getToken(String code);
