+++ src/main/java/com/nhnacademy/authservice/controller/AuthController.java
@@ -1,6 +1,8 @@
 package com.nhnacademy.authservice.controller;
 
-import com.nhnacademy.authservice.domain.*;
+import com.nhnacademy.authservice.domain.request.LoginRequestDto;
+import com.nhnacademy.authservice.domain.response.RefreshTokenResponseDto;
+import com.nhnacademy.authservice.domain.response.TokenParseResponseDto;
 import com.nhnacademy.authservice.service.auth.AuthService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
