--- src/main/java/com/nhnacademy/authservice/domain/UserUpdateRequest.java
@@ -1,16 +1,0 @@
-package com.nhnacademy.authservice.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-@AllArgsConstructor
-public class UserUpdateRequest {
-    private String userPassword;
-    private String userName;
-    private String userPhoneNumber;
-    private String userEmail;
-    private LocalDate userBirth;
-}
