+++ src/main/java/com/nhnacademy/authservice/domain/request/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.authservice.domain;
+package com.nhnacademy.authservice.domain.request;
 
 public record LoginRequestDto(
         String id,
