+++ src/main/java/com/nhnacademy/authservice/domain/request/OAuth2LoginRequestDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.authservice.domain;
+package com.nhnacademy.authservice.domain.request;
 
 import lombok.Data;
