+++ src/main/java/com/nhnacademy/authservice/service/auth/AuthServiceImpl.java
@@ -3,7 +3,9 @@ package com.nhnacademy.authservice.service.auth;
 import com.nhnacademy.authservice.adapter.UserAdapter;
 import com.nhnacademy.authservice.client.member.OAuth2MemberClient;
 import com.nhnacademy.authservice.client.token.OAuth2TokenClient;
-import com.nhnacademy.authservice.domain.*;
+import com.nhnacademy.authservice.domain.request.OAuth2AdditionalSignupRequestDto;
+import com.nhnacademy.authservice.domain.request.OAuth2UserCreateRequestDto;
+import com.nhnacademy.authservice.domain.response.*;
 import com.nhnacademy.authservice.exception.InvalidTokenException;
 import com.nhnacademy.authservice.factory.OAuth2MemberClientFactory;
 import com.nhnacademy.authservice.factory.OAuth2TokenClientFactory;
