+++ src/test/java/com/nhnacademy/authservice/userdetails/CustomUserDetailsTest.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.authservice.userdetails;
 
-import com.nhnacademy.authservice.domain.UserResponse;
+import com.nhnacademy.authservice.domain.response.UserResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.security.core.GrantedAuthority;
